Add web interface for Dots and Boxes game with Flask
A new web interface for the Dots and Boxes game is added using Flask. This feature includes a 'click' method that handles game logic whenever a player makes a move on the game board. A few UI adjustments have been made for the Tkinter application to work with Flask without affecting the GUI. A set of unit tests is also provided to validate the new interface.